perf(refine): AVX2 z-normalize — f64 moments + f32 normalize
cache_znorm was 17.4% of refine_total (the next phase after consensus). Vectorize
znormalize_into: weighted_moments (the Σw·x / Σw·x² reduction) as a 4-wide AVX2
f64 kernel, and znorm_write (the per-pixel out = √w·(x−mean)·inv_norm) as an
8-wide AVX2 f32 kernel; scalar fallbacks + n%K tails + parity tests.

The moments stay f64 on purpose: the z-normalized stack is the consensus input
(unlike the IRLS residual, which only feeds the reweight), so the
cancellation-prone variance s2 − s1·mean feeds Φ directly. A measured full-f32
moments path moved 2.3% of normals (max 30°, max|ΔΦ| 0.06); f64 moments keeps
the win accuracy-neutral. The f32 √w for the normalize kernel is computed once
per level (in coarse_to_fine) and threaded through, so the hot cache path does
no per-candidate weight conversion.

Measured (big dino, 56,542 pts, R24 fronto cache, robust3, 4 cores):
- cache_znorm 8.12 -> 3.37 us/call (2.4x; hoisting the √w conversion trims a bit
  more on top)
- normals preserved: median/p99 0.000 deg, 0.0% move >0.01 deg (one lone
  ambiguous point flips 0.93 deg, Φ-equivalent), max|ΔΦ| 4.6e-4
- full-cloud wall ~70s (~1.1x e2e by Amdahl on the 17.4% share; absolute walls
  carry VM load noise, so the 2.4x phase number is the reliable metric)

+++ crates/sfmtool-core/src/patch_normal_refine/fronto_cache.rs
@@ -369,7 +369,7 @@ pub(super) fn eval_phi(
     resolution: u32,
     cols: &[i32],
     rows: &[i32],
-    sqrt_w: &[f64],
+    sqrt_w: &[f32],
     total_w: f64,
     scratch: &mut Scratch,
     objective: Objective,
